Add forms: rename checkout file to forms, add more different inputs, use mask plugin for input
