+++ package.json
@@ -51,6 +51,7 @@
     "@babel/polyfill": "^7.12.1",
     "jquery": "^3.6.0",
     "jquery-mask-plugin": "^1.14.16",
+    "js-datepicker": "^5.18.0",
     "normalize.css": "^8.0.1",
     "slick-carousel": "^1.8.1"
   }
