+++ src/styles/assets/_colors.scss
@@ -3,6 +3,7 @@ $white-border: #f2f2f2;
 $white-opacity2: rgba(255, 255, 255, 0.2);
 $white-opacity1: rgba(255, 255, 255, 0.2);
 $black: #000000;
+$black-bg-modal: rgba(0, 0, 0, 0.5);
 $black-title: #111111;
 $black-text: #333333;
 $black-bg: #2c332c;
