+++ src/styles/elements/_login.scss
@@ -1,0 +1,78 @@
+.login {
+  &__window {
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    right: 0;
+    left: 0;
+    background-color: $black-bg-modal;
+    z-index: 99;
+    opacity: 0;
+    pointer-events: none;
+    transition: opacity 0.2s;
+  }
+
+  &__page {
+    width: 30%;
+    padding: 50px 30px 20px;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    border-radius: 10px;
+    transform: translate(100%, -50%);
+    background: $white;
+    transition: transform cubic-bezier(0.68, -0.55, 0.265, 1.55) 0.3s;
+
+    @include md {
+      width: 95%;
+    }
+  }
+  &__form {
+    display: grid;
+    &-title {
+      margin-bottom: 20px;
+    }
+
+    & input,
+    & textarea {
+      background-color: $grey-input-bg;
+      padding: 12px 20px;
+      border: 0;
+      outline: none;
+
+      &::placeholder {
+        font-family: $jost;
+        font-weight: normal;
+        font-size: 14px;
+        line-height: 22px;
+        color: $grey;
+      }
+    }
+  }
+
+  &__text {
+    margin-top: 20px;
+  }
+  &__link {
+    color: $green;
+  }
+
+  &__button {
+    &_close {
+      position: absolute;
+      top: 50px;
+      right: 30px;
+    }
+  }
+}
+
+.login_open {
+  & .login__page {
+    transform: translate(-50%, -50%);
+  }
+
+  &.login__window {
+    pointer-events: initial;
+    opacity: 1;
+  }
+}
