+++ src/styles/sections/_checkout.scss
@@ -1,21 +1,26 @@
 .checkout {
   padding-top: 80px;
   padding-bottom: 80px;
-  display: flex;
+  //   display: flex;
+  display: grid;
+  grid-template-columns: 1fr auto;
 
   @include md {
     padding-top: 20px;
     padding-bottom: 20px;
   }
 
-  @include md {
-    // flex-direction: column;
+  //   @include md {
+  //     // flex-direction: column;
 
-    display: grid;
-  }
+  //     display: grid;
+  //   }
 
   &__link {
     color: $green;
+    &:hover {
+      cursor: pointer;
+    }
   }
 
   &__form {
@@ -33,11 +38,6 @@
     }
 
     &-title {
-      font-family: $jost;
-      font-weight: 600;
-      font-size: 18px;
-      line-height: 22px;
-      color: $black;
       margin-bottom: 30px;
 
       &:not(:first-of-type) {
@@ -45,11 +45,25 @@
       }
     }
 
-    & input {
+    & input,
+    & textarea,
+    & select {
       background-color: $grey-input-bg;
       padding: 12px 20px;
       border: 0;
+      border-left: 5px solid transparent;
       outline: none;
+      font-family: $jost;
+
+      font-size: 16px;
+      line-height: 22px;
+      color: $black-text;
+
+      &:focus,
+      &:active {
+        background-color: darken($color: $grey-input-bg, $amount: 5);
+        border-left: 5px solid $grey;
+      }
 
       &::placeholder {
         font-family: $jost;
@@ -71,9 +85,11 @@
     font-weight: 400;
     font-size: 15px;
     line-height: 36px;
+    grid-column: 1 / 3;
 
     @include md {
       font-size: 14px;
+      grid-column: 1 /2;
     }
   }
 
@@ -114,6 +130,20 @@
       right: 10px;
     }
   }
+
+  &__button {
+    @include md {
+      grid-column: 1 /3;
+    }
+
+    &_wrapper {
+      grid-column: 1 /3;
+      display: grid;
+      grid-template-columns: 1fr 1fr;
+      gap: 20px;
+      margin-top: 20px;
+    }
+  }
 }
 
 .return_link {
@@ -127,6 +157,14 @@
   & span {
     font-family: $icons;
   }
+}
+
+.form-title {
+  font-family: $jost;
+  font-weight: 600;
+  font-size: 18px;
+  line-height: 22px;
+  color: $black;
 }
 
 .form-text {
@@ -201,4 +239,21 @@
   &__adress {
     margin-bottom: 5px;
   }
+}
+
+.amount {
+  text-align: right;
+}
+
+.input {
+  &__color {
+    padding: 0;
+    height: 50px;
+    width: 100%;
+  }
+}
+
+label {
+  font-family: $jost;
+  font-size: 16px;
 }
