+++ src/styles/styles.scss
@@ -1,4 +1,5 @@
 @import "../../node_modules/normalize.css";
+@import "~js-datepicker/src/datepicker";
 @import "assets/colors";
 @import "assets/fonts";
 @import "assets/mixins";
@@ -11,10 +12,12 @@
 @import "elements/content_box";
 @import "elements/copyright";
 @import "elements/dropdown_content";
+@import "elements/gift-card";
 @import "elements/hover";
 @import "elements/icon";
 @import "elements/item";
 @import "elements/link";
+@import "elements/login";
 @import "elements/menu";
 @import "elements/nav";
 @import "elements/reserve";
